Break legend-dialog import cycle and pin test locale
